changed concatIterate to use stack of stream frames
This seems to have solved the overhead that was initially added by the
concatIterate approach. I was initially thinking that the Semigroup
instances for Stream and StreamFrame could adopt the state sequence
approach, but in practice I'm not sure if it could alleviate the deep
nesting issue due to the fact that you'd still be building up a tree of
states instead of a sequence of them.