initial attempt at parallelizing dir traversal
After an intense struggle with the ins and outs of haskell concurrency,
foreign function calls, and the posix api, the perf here is still
slightly worse than the single-threaded perf (lolololol). The profiler
says that the priority queue is by far the largest time hog, which is
unfortunate since it kinda underpins the whole system of parallel
directory traversal with output order maintained.

The next step will probably be to swap out the priority queue for a
regular one just to verify that that's the bottleneck. If it is the
bottleneck, then uhhhhhhhhh guess I gotta dive into priority queues and
figure out what's going on there perf-wise.