switched to bound threads to prevent dir calls from blocking other shiz
This did not increase performance at the moment, but that's moreso
because the bottleneck is the priority queue. But either way we still
needed to make this change once we fix this bottleneck.

And speaking of the bottleneck, after having a look at the fd repo (with
a lot of help from deepwiki), apparently they don't use a priority queue
at all. Instead, any run of fd that results in 1000 or fewer results
will simply sort the results and then output them. And if 1000 results
is exceeded (or 100ms is exceeded), then fd doesn't even attempt to keep
the results ordered and instead just prints shit as it comes in. Also,
the worker threads batch 256 results before sending them to the consumer
to reduce synchronization overhead.

Sooooooooo basically I need to ditch the priority queue and only bother
with ordered results if there's less than 1000 of them. Also need the
workers to batch results so that we're not slamming the shared queues.