cut thread count in half if there's no regex matching
The idea here is the threads will be mostly IO bound if there's no regex
matching and as such won't benefit from the default number of threads.
This does increase the performance of cached runs but decreases the
performance of non-cached runs, so it's unclear whether or not this is a
net positive change.