initial implementation of glob parser/matcher
When I first implemented the -g flag, I did so going purely off of what
the fd readme says about it rather than thinking for 2 seconds and
realizing that the glob flag does indeed do glob matches and not just
exact string matches. So this commit corrects that bozo move.

Unfortunately the performance of the glob matcher sucks when using a
glob with `*` in it. Will need to look up how to make that more
performant.